[tests][xtro] Add enum member/values checks. Fix #7527 (#9691)
Figure out if
* we're missing enum values (easy to workaround, but annoying for developers)
* we expose enum values that are not defined natively (potential bugs)

reference: https://github.com/xamarin/xamarin-macios/issues/7527

+++ src/CoreWlan/Enums.cs
@@ -83,7 +83,7 @@ namespace CoreWlan {
 		Wpa3Enterprise = 12,
 		[Mac (10,15)]
 		Wpa3Transition = 13,
-		Unknown = int.MaxValue,
+		Unknown = long.MaxValue,
 	}
 
 	[Deprecated (PlatformName.MacOSX, 11,0)]
